Client login on cron time
The problem with the bot before was that the client session would expire after several hours, so by the time the scheduled cron time rolled around an error would be thrown. Now the client is created at the time specified, so the error will not occur no matter how far in the future the time is.